Simply simple examples (#258)
the mnist-convnet.py examples was overloaded. This commit splits the
file into the most minimal mnist-convnet and mnist-tfslim. Further
it adds an example about the visualization of filters which
illustrates the usage of `VariableHolder`.

In addition, this also adds a boilerplate template for projects,
which might facilitate the process of starting a new project.