Clearify LMDB-dumps and MinSaver in Documentation
The user might be familiar with creating a ds-stream and wants to
dump the compressed image to disk as an lmdb file. This adds this to
the documentation but gives a warning about the write-performance.

In addition, there is a small amount of  documentation about the
Min-Saver and ScalarStats-runner, which is slightly now described in
more detail.