Atari Visualization Bugfix (#529)
* Atari Visualization Bugfix

OpenCV2 does not always display the visualization properly. 

For imshow to properly update, cv2.waitKey should be called. Without this line, the cv2 window does not refresh on my system with OpenCV version 2.4.9.1 on Debian with Python 2.7.9. More info can be found on [StackOverflow](https://stackoverflow.com/questions/44854699/why-does-the-cv2-imshow-does-not-render-without-cv2-waitkey).

* Remove unused import